+++ src/main/java/net/silentchaos512/loginar/compat/SgearCompatProxy.java
@@ -1,24 +1,14 @@
 package net.silentchaos512.loginar.compat;
 
-import net.minecraft.resources.ResourceLocation;
-import net.minecraft.tags.ItemTags;
-import net.minecraft.tags.TagKey;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.silentchaos512.lib.util.Color;
 
 import java.util.Optional;
 
 final class SgearCompatProxy {
-    private static final TagKey<Item> BORT_TAG = ItemTags.create(ResourceLocation.fromNamespaceAndPath("c", "gems/bort"));
     private SgearCompatProxy() {}
 
     static Optional<Color> getMainPartColor(ItemStack stack) {
-        // Bort exception -- TODO: This should be in the main color map
-        if (stack.is(BORT_TAG)) {
-            return Optional.of(new Color(0x96A3D4));
-        }
-
         // Check material color
         /*MaterialInstance material = MaterialInstance.from(stack);
         if (material != null) {